// web/src/api/models/search-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { FulltextSearchParams } from './fulltext-search-params';
// May contain unused imports in some cases
// @ts-ignore
import type { GraphSearchParams } from './graph-search-params';
// May contain unused imports in some cases
// @ts-ignore
import type { SummarySearchParams } from './summary-search-params';
// May contain unused imports in some cases
// @ts-ignore
import type { VectorSearchParams } from './vector-search-params';
// May contain unused imports in some cases
// @ts-ignore
import type { VisionSearchParams } from './vision-search-params';

/**
 * Search request
 * @export
 * @interface SearchRequest
 */
export interface SearchRequest {
    /**
     * 
     * @type {string}
     * @memberof SearchRequest
     */
    'query'?: string;
    /**
     * 
     * @type {VectorSearchParams}
     * @memberof SearchRequest
     */
    'vector_search'?: VectorSearchParams;
    /**
     * 
     * @type {FulltextSearchParams}
     * @memberof SearchRequest
     */
    'fulltext_search'?: FulltextSearchParams;
    /**
     * 
     * @type {GraphSearchParams}
     * @memberof SearchRequest
     */
    'graph_search'?: GraphSearchParams;
    /**
     * 
     * @type {SummarySearchParams}
     * @memberof SearchRequest
     */
    'summary_search'?: SummarySearchParams;
    /**
     * 
     * @type {VisionSearchParams}
     * @memberof SearchRequest
     */
    'vision_search'?: VisionSearchParams;
    /**
     * Whether to save search result to database history
     * @type {boolean}
     * @memberof SearchRequest
     */
    'save_to_history'?: boolean;
    /**
     * Whether to enable rerank for search results
     * @type {boolean}
     * @memberof SearchRequest
     */
    'rerank'?: boolean;
}